Updates to phoronix-install-playbook.yml
Still has problems with connection to github timing out on a seemingly random one of the remote while cloning the phoronix repo. Other than that, it succeeds on most of the remote hosts each time and RAMspeed seems to be running successfully on the remote machines when run on them locally and manually.